api function addition
